requirements.txt
tkintertable
pyshorteners
clipboard
requests
qrcode
Pillow